fix(scoring): move pure functions out of server action file to fix build
